1.) fixed a memory leak issue by centralizing all the places I put I use 'new' into the memory class 2.) fixed issues that prevented recursion 3.) fixed issues with multiple returns in a function, now returns can be used anywhere as many times as desired in a function

+++ src/Types/Number/Derived/Int.cpp
@@ -24,12 +24,12 @@ int Int::Set(int64 &data) {
 }
 
 int Int::Set(float &data) {
-    m_data.integer = (int)data;
+    m_data.integer = data;
     return SUCCEED;
 }
 
 int Int::Set(long double &data) {
-    m_data.integer = (int)(float)data;
+    m_data.integer = data;
     return SUCCEED;
 }
 
@@ -45,23 +45,23 @@ int Int::Get(int64 &data) {
 }
 
 int Int::Add(Number &other) {
-    m_data.integer += (int)(float)GetOtherNumber(other);
+    m_data.integer += GetOtherNumber(other);
     return SUCCEED;
 }
 
 int Int::Subtract(Number &other) {
-    m_data.integer -= (int)(float)GetOtherNumber(other);
+    m_data.integer -= GetOtherNumber(other);
     return SUCCEED;
 }
 
 int Int::Multiply(Number &other) {
-    m_data.integer *= (int)(float)GetOtherNumber(other);
+    m_data.integer *= GetOtherNumber(other);
     return SUCCEED;
 }
 
 int Int::Divide(Number &other) {
     try {
-        m_data.integer /= (int)(float)DivideByZeroGuard(other);
+        m_data.integer /= DivideByZeroGuard(other);
     }
     catch(const char* msg){
         cout << msg;
